remove amp tests
