+++ cypress/e2e/pages/onDemandAudio/index.cy.js
@@ -1,9 +1,7 @@
 import config from '../../../support/config/services';
 import getPaths from '../../../support/helpers/getPaths';
 import serviceHasPageType from '../../../support/helpers/serviceHasPageType';
-import testsForCanonicalOnly from './testsForCanonicalOnly';
-import testsForAMPOnly from './testsForAMPOnly';
-import crossPlatformTests from './tests';
+import e2eTests from './tests';
 import visitPage from '../../../support/helpers/visitPage';
 import { overrideRendererOnTest } from '../../../support/helpers/onDemandRadioTv';
 
@@ -24,37 +22,11 @@ Object.keys(config)
 
           visitPage(newPath, pageType);
         });
-        crossPlatformTests({
-          service,
-          pageType,
-          variant,
-        });
-        testsForCanonicalOnly({
+        e2eTests({
           service,
           pageType,
           variant,
         });
       });
     });
-    paths
-      .map(path => `${path}.amp`)
-      .forEach(currentPath => {
-        describe(`${pageType} - ${currentPath}`, () => {
-          before(() => {
-            Cypress.env('currentPath', currentPath);
-            visitPage(currentPath, pageType);
-          });
-          crossPlatformTests({
-            service,
-            pageType,
-            variant,
-            isAmp: true,
-          });
-          testsForAMPOnly({
-            service,
-            pageType,
-            variant,
-          });
-        });
-      });
   });
