+++ src/app/pages/OnDemandAudioPage/index.stories.jsx
@@ -30,7 +30,6 @@ const Component = (_, { service }) => {
         pageData={onDemandRadioFixtures[service]}
         status={200}
         service={service}
-        isAmp={false}
         loading={false}
         error=""
         pageType={MEDIA_PAGE}
