--- src/integration/pages/onDemandAudioPage/ampTests.js
@@ -1,13 +1,0 @@
-import {
-  runCoreAmpTests,
-  runAmpAnalyticsTests,
-  runRadioScheduleTests,
-  runSeoAmpTests,
-} from '../../common';
-
-export default () => {
-  runCoreAmpTests();
-  runAmpAnalyticsTests();
-  runRadioScheduleTests({ isAmp: true });
-  runSeoAmpTests();
-};
