--- src/integration/pages/onDemandAudioPage/canonicalTests.js
@@ -1,11 +1,0 @@
-import {
-  runCoreCanonicalTests,
-  runCanonicalAnalyticsTests,
-  runRadioScheduleTests,
-} from '../../common';
-
-export default () => {
-  runCoreCanonicalTests();
-  runCanonicalAnalyticsTests();
-  runRadioScheduleTests({ isAmp: false });
-};
