+++ src/integration/pages/onDemandAudioPage/crossPlatformTests.js
@@ -1,4 +1,7 @@
 import {
+  runCoreCanonicalTests,
+  runCanonicalAnalyticsTests,
+  runRadioScheduleTests,
   runCommonCrossPlatformTests,
   runMediaPlayerEmbedLegacyTests,
   runRecentEpisodesTests,
@@ -10,4 +13,7 @@ export default service => {
   runCommonEpisodeTests();
   runRecentEpisodesTests();
   runMediaPlayerEmbedLegacyTests();
+  runCoreCanonicalTests();
+  runCanonicalAnalyticsTests();
+  runRadioScheduleTests({ isAmp: false }); // keep this for now until live radio also has amp removed
 };
