--- src/integration/pages/onDemandAudioPage/indonesia.expired/amp.test.js
@@ -1,14 +1,0 @@
-/**
- * @service indonesia
- * @pathname /indonesia/bbc_indonesian_radio/w172x6r5000f38s
- */
-
-import runExpiredEpisodeTests from '../expiredEpisodeTests';
-import runAmpTests from '../ampTests';
-
-describe('AMP', () => {
-  describe(pageType, () => {
-    runExpiredEpisodeTests();
-    runAmpTests();
-  });
-});
