+++ src/integration/pages/onDemandAudioPage/indonesia.expired/canonical.test.js
@@ -2,13 +2,24 @@
  * @service indonesia
  * @pathname /indonesia/bbc_indonesian_radio/w172x6r5000f38s
  */
-
+import {
+  runCoreCanonicalTests,
+  runCanonicalAnalyticsTests,
+  runRadioScheduleTests,
+  runCommonCrossPlatformTests,
+  runRecentEpisodesTests,
+} from '../../../common';
+import runCommonEpisodeTests from '../commonEpisodeTests';
 import runExpiredEpisodeTests from '../expiredEpisodeTests';
-import runCanonicalTests from '../canonicalTests';
 
 describe('Canonical', () => {
   describe(pageType, () => {
     runExpiredEpisodeTests();
-    runCanonicalTests();
+    runCommonCrossPlatformTests(service);
+    runCommonEpisodeTests();
+    runRecentEpisodesTests();
+    runCoreCanonicalTests();
+    runCanonicalAnalyticsTests();
+    runRadioScheduleTests({ isAmp: false });
   });
 });
