--- src/integration/pages/onDemandAudioPage/indonesia/amp.test.js
@@ -1,14 +1,0 @@
-/**
- * @service indonesia
- * @pathname /indonesia/bbc_indonesian_radio/w172xh267fpn19l
- */
-
-import runCrossPlatformTests from '../crossPlatformTests';
-import runAmpTests from '../ampTests';
-
-describe('AMP', () => {
-  describe(pageType, () => {
-    runCrossPlatformTests(service);
-    runAmpTests();
-  });
-});
