+++ src/integration/pages/onDemandAudioPage/indonesia/canonical.test.js
@@ -4,11 +4,9 @@
  */
 
 import runCrossPlatformTests from '../crossPlatformTests';
-import runCanonicalTests from '../canonicalTests';
 
 describe('Canonical', () => {
   describe(pageType, () => {
     runCrossPlatformTests(service);
-    runCanonicalTests();
   });
 });
