+++ src/integration/pages/onDemandAudioPage/pashto.expired/canonical.test.js
@@ -2,13 +2,24 @@
  * @service pashto
  * @pathname /pashto/bbc_pashto_radio/w172x8nvf4bchz5
  */
-
+import {
+  runCoreCanonicalTests,
+  runCanonicalAnalyticsTests,
+  runRadioScheduleTests,
+  runCommonCrossPlatformTests,
+  runRecentEpisodesTests,
+} from '../../../common';
+import runCommonEpisodeTests from '../commonEpisodeTests';
 import runExpiredEpisodeTests from '../expiredEpisodeTests';
-import runCanonicalTests from '../canonicalTests';
 
 describe('Canonical', () => {
   describe(pageType, () => {
     runExpiredEpisodeTests();
-    runCanonicalTests();
+    runCommonCrossPlatformTests(service);
+    runCommonEpisodeTests();
+    runRecentEpisodesTests();
+    runCoreCanonicalTests();
+    runCanonicalAnalyticsTests();
+    runRadioScheduleTests({ isAmp: false });
   });
 });
