--- src/integration/pages/onDemandAudioPage/pashto/amp.test.js
@@ -1,14 +1,0 @@
-/**
- * @service pashto
- * @pathname /pashto/bbc_pashto_radio/w3ct0lz1
- */
-
-import runCrossPlatformTests from '../crossPlatformTests';
-import runAmpTests from '../ampTests';
-
-describe('AMP', () => {
-  describe(pageType, () => {
-    runCrossPlatformTests(service);
-    runAmpTests();
-  });
-});
